refactor: migrate to myUser resolver and optimize fetched data on site load
Merged in LT-53-redeuce-queries (pull request #701)

LT-53 reduce queries

* test: refactor and improve tests

* Merge branch 'dev' into LT-53-redeuce-queries

# Conflicts:
#	src/api/organizationMemberships.ts
#	src/store/useDashboardMode.ts

* fix: types and remove unused files and functions

* feat: remove jest-svg-transformer

* chore: remove commented code and update .env.example

* fix: reuse existing RoleConnectionNode interface

* fix: add currentOrganization.id guards

* fix: rename roleName array variables and fix type issues

* fix: possible memory leak when unmounting assessments graph/table

* Merge branch 'dev' into LT-53-redeuce-queries

# Conflicts:
#	src/components/Organization/JoinedOrganizationTable.test.tsx
#	src/components/Organization/JoinedOrganizationTable.tsx
#	src/components/User/Dialog/Edit.test.tsx
#	src/components/User/Dialog/Edi…

Approved-by: Ricardo Barata
Approved-by: Malcolm McConaghy

+++ src/components/User/Dialog/Edit.tsx
@@ -86,11 +86,11 @@ export default function EditUserDialog (props: Props) {
     const [ formErrors, setFormErrors ] = useState<Errors>({});
     const [ valid, setValid ] = useState(true);
     const currentOrganization = useCurrentOrganization();
-    const organizationId = currentOrganization?.organization_id ?? ``;
+    const organizationId = currentOrganization?.id ?? ``;
     const { data: organizationMembershipData, loading: loadingMembershipData } = useGetOrganizationUserNode({
         fetchPolicy: `cache-and-network`,
         variables: {
-            id: userId ?? ``,
+            userId: userId ?? ``,
             organizationId,
         },
         skip: !open || !organizationId || !userId,
@@ -170,7 +170,7 @@ export default function EditUserDialog (props: Props) {
     };
 
     const handleDelete = async () => {
-        if (!(userId && organizationId)) return;
+        if (!userId || !organizationId) return;
         const { givenName, familyName } = organizationMembershipData?.userNode ?? {};
         const userName = `${givenName} ${familyName}`;
         if (!(await deletePrompt({
